Adding waypoint at end wraps to beginning
There was an issue that way far more visible in the parking lot scene - where slicing waypoints to publish, where the end of the slice was larger than our waypoints list would clamp to the end. Now if we have extra waypoints that we could still publish when we hit the end, we wrap around and add beginning waypoints.